March 2025 release

+++ preprocessors/preprocessor_template/h_[name_of_preprocessor].h
@@ -1,4 +1,4 @@
-static inline void h_[name_of_preprocessor](int& size, byte*& data, const int paramc, const double paramv [])  // CPU preprocessor encoder
+static inline void h_[name_of_preprocessor](long long& size, byte*& data, const int paramc, const double paramv [])  // CPU preprocessor encoder
 {
   // transforms the 'size' bytes in the 'data' array and writes the result either back to the 'data' array or to a new array and then makes 'data' point to this new array
   // if the number of bytes changes, the 'size' needs to be updated accordingly
@@ -8,7 +8,7 @@ static inline void h_[name_of_preprocessor](int& size, byte*& data, const int pa
   // this code must be manually parallelized (using OpenMP) if desired
 }
 
-static inline void h_i[name_of_preprocessor](int& size, byte*& data, const int paramc, const double paramv [])  // CPU preprocessor decoder
+static inline void h_i[name_of_preprocessor](long long& size, byte*& data, const int paramc, const double paramv [])  // CPU preprocessor decoder
 {
   // transforms the 'size' bytes in the 'data' array and writes the result either back to the 'data' array or to a new array and then makes 'data' point to this new array
   // if the number of bytes changes, the 'size' needs to be updated accordingly
